refactor: dotenv 위치 및 types 모듈화
Related to: #143
